refactor(web): give the tools popover an agent it can rely on
The popover resolved its own agent and then had to ask whether it had one
at sixteen places, because `useActiveAgent` returns undefined and every
row, toggle and source in the panel is scoped to an agent.

It splits in two. The exported component resolves the agent and decides
what no agent means; the inner one takes a required agent and never asks.
The body moves across unchanged.

No agent covers two situations, neither of which wants a tool menu: the
agent list is still loading, which is every cold mount, or there is
genuinely no agent to act as — the default Assistant disabled with nothing
to replace it — which AppPage already answers with NoAgentModal.

Keyed on the agent, so switching starts clean rather than carrying the
previous agent's open panel and search term across.

Renamed from ActionsPopover: it lists tools, and the directory it now
lives in is lib/tools. The e2e page object follows.
